move transport methods to fleetscheduler simplify workerfactory generalize and move transports to own settings section fix negative payload
