--- TBot/Workers/Brain/IAutoMineWorker.cs
@@ -1,13 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TBot.Ogame.Infrastructure.Enums;
-using TBot.Ogame.Infrastructure.Models;
-
-namespace Tbot.Workers.Brain {
-	public interface IAutoMineWorker {
-		Task<int> HandleMinerTransport(Celestial origin, Celestial destination, Resources resources, Buildables buildable = Buildables.Null, Buildings maxBuildings = null, Facilities maxFacilities = null, Facilities maxLunarFacilities = null, AutoMinerSettings autoMinerSettings = null);
-	}
-}
